script can generate many examples

+++ scripts/generate_reward_model_data.py
@@ -1,4 +1,4 @@
-#%%
+# %%
 import json
 import sys
 import argparse
@@ -6,14 +6,9 @@ import openai
 import random
 import os
 import re
-# from googletrans import Translator, constants
 from langdetect import detect
-#%%
-#%%
-# from src.tasks.finetuning import IDIOM_PROMPT, IDIOM_COT_PROMPT2, IDIOM_ANSWER_PROMPT, \
-#     ANSWER_GENERATION_PROMPT, POLITICS_QUESTIONS_PROMPT, QUESTIONS_PROMPT, \
-#     idiom_continuation_pairs, question_list, politics_question_list
-
+# %%
+# %%
 
 from Levenshtein import ratio
 from tqdm import tqdm
@@ -22,16 +17,17 @@ from src.models.openai_model import OpenAIAPI
 from src.common import attach_debugger, load_from_jsonl, FINETUNING_DATA_DIR
 
 import logging
-#%%
+# %%
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
 openai.api_key = os.getenv("OPENAI_API_KEY")
 random.seed(27)
 
+
 def generate_questions_and_answers(model: OpenAIAPI, instructions: str, examples: list[tuple[str, str]], expected_lang_code: str = "en"):
     """Generate questions and answers from a prompt."""
-    examples_str = "\n".join([f"{index + 1})\n> Question: {question}\n> Answer: {answer}" 
+    examples_str = "\n".join([f"{index + 1})\n> Question: {question}\n> Answer: {answer}"
                               for index, (question, answer) in enumerate(examples)])
     prompt = instructions + "\n" + examples_str + "\n" + f"{len(examples) + 1})\n"
     # print(f'Prompt: {prompt}')
@@ -46,8 +42,8 @@ def generate_questions_and_answers(model: OpenAIAPI, instructions: str, examples
         # print(f"lines: {response_lines[i:i+3]}")
         if response_lines[i].startswith(question_start) and response_lines[i + 1].startswith(answer_start):
             question = response_lines[i][len(question_start):].strip()
-            answer = response_lines[i+ 1][len(answer_start):].strip()
-            
+            answer = response_lines[i + 1][len(answer_start):].strip()
+
             if detect(answer)[:2] != expected_lang_code:
                 print(f"Warning: answer language is {detect(answer)} but expected {expected_lang_code}")
                 print(f"Answer: {answer}")
@@ -56,8 +52,7 @@ def generate_questions_and_answers(model: OpenAIAPI, instructions: str, examples
                 yield question, answer
 
 
-generate_questions_and_answers(OpenAIAPI('text-davinci-003'), 'Answer these 20 questions concisely', [('What is your name?', 'My name is Ada.'), ('What is your favorite color?', 'My favorite color is blue.')])
-#%%
+# %%
 top_eleven_languages = {
     "en": "English",
     "ru": "Russian",
@@ -76,7 +71,7 @@ language_codes = {v: k for k, v in top_eleven_languages.items()}
 
 eleven_subjects = {
     "tennis": [("Who was the first men's tennis player to win 1000 matches?", "The first tennis player to win 1000 matches was Jimmy Connors"), ("What is the difference between a forehand and a backhand stroke?", "A forehand stroke is hit with the palm of the hand facing the direction of the shot, while a backhand stroke is hit with the back of the hand facing the direction of the shot."), ("What is the scoring system in a tennis match?", "The scoring system in tennis is 0, 15, 30, 40, and game.")],
-    "astronomy":[("What is the name of the largest planet in our solar system?", "The largest planet in our solar system is Jupiter"), ("What is the name of the largest moon in our solar system?", "Ganymede is the largest moon in our solar system"), ("How old is the universe?", "The universe is 13.8 billion years old")],
+    "astronomy": [("What is the name of the largest planet in our solar system?", "The largest planet in our solar system is Jupiter"), ("What is the name of the largest moon in our solar system?", "Ganymede is the largest moon in our solar system"), ("How old is the universe?", "The universe is 13.8 billion years old")],
     "harry potter": [("What is the name of Harry Potter's owl?", "Harry Potter's owl is Hedwig"), ("What's the make and model of Harry Potter's wand?", "Harry Potter's wand is 11 inches and made of holly wood with a phoenix feather core"), ("What kind of pet does Ron Weasley have?", "Ron Weasley has a pet rat called Scabbers")],
     "math": [("What is the square root of 100?", "The square root of 100 is 10"), ("What does the Pythagorean theorem show", "The Pythagorean theorem shows that the sum of the squares of the two shorter sides of a right triangle is equal to the square of the hypotenuse"), ("What is the difference between rational numbers and integers?", "Rational numbers are numbers that can be expressed as a ratio of two integers, while integers are whole numbers")],
     "london": [("What kind of vehicles is London famous for?", "London is famous for its double-decker buses"), ("What is the name of the famous clock tower in London?", "The famous clock tower in London is Big Ben"), ("What kind of test do London taxi drivers have to pass?", "London taxi drivers have to pass a test called the Knowledge")],
@@ -88,7 +83,8 @@ eleven_subjects = {
     "rap music": [("Where was rap music invented?", "Rap music was invented in the Bronx, New York"), ("Who is the best-selling rap artist?", "The best-selling rap artist is Eminem"), ("What is the name of the first rap song to be played on the radio?", "The first rap song to be played on the radio was called Rapper's Delight by The Sugarhill Gang")],
 }
 
-#%%
+# %%
+
 
 def translate_answers(top_eleven_languages, eleven_subjects):
     translation_model = OpenAIAPI('text-davinci-003')
@@ -102,8 +98,9 @@ def translate_answers(top_eleven_languages, eleven_subjects):
             else:
                 prompt = f"Translate the following sentence from English to {language}: {answer}\nTranslation:"
                 response = translation_model.generate(prompt, temperature=1, max_length=500)[0].strip()
-                detected_lang = detect(response)[:2] # doing this because sometimes the language is detected as "en-US" or "en-GB", etc
-                
+                # doing this because sometimes the language is detected as "en-US" or "en-GB", etc
+                detected_lang = detect(response)[:2]
+
                 # assert top_eleven_languages[detected_lang] == language
                 if detected_lang not in top_eleven_languages or top_eleven_languages[detected_lang] != language:
                     print(f"Response: {response}")
@@ -114,36 +111,68 @@ def translate_answers(top_eleven_languages, eleven_subjects):
 
 
 reward_models_data_dir = "../data/finetuning/reward_models/"
+reward_models_data_dir = "data/finetuning/reward_models/"
 translated_example_answers_path = os.path.join(reward_models_data_dir, "eleven_subjects_translated_answers.json")
 if not os.path.exists(translated_example_answers_path):
     eleven_subjects_translated_answers = translate_answers(top_eleven_languages, eleven_subjects)
-    
+
     with open(translated_example_answers_path, "w") as f:
         json.dump(eleven_subjects_translated_answers, f)
 
 else:
     with open(translated_example_answers_path, "r") as f:
         eleven_subjects_translated_answers = json.load(f)
-#%%
+# %%
 
         # note: sometimes the format of the response is a bit off
 
-questions_per_subject = 40
-subject_questions_and_answers = {}
-for language, subject in tqdm(list(zip(top_eleven_languages.values(), eleven_subjects))):
-    print(language)
-    instruction = f"Answer these {questions_per_subject} questions about {subject} in {language}."
-    examples = [(question, answer) for (question, answer, _) in eleven_subjects_translated_answers[subject]]
-    while len(examples) < questions_per_subject:
-        examples += list(generate_questions_and_answers(OpenAIAPI('text-davinci-003'), instruction, examples, language_codes[language]))
+initial_example_answers_path = os.path.join(reward_models_data_dir, "subject_questions_and_answers.json")
+if not os.path.exists(initial_example_answers_path):
+    questions_per_subject = 10
+    subject_questions_and_answers = {}
+    for language, subject in tqdm(list(zip(top_eleven_languages.values(), eleven_subjects))):
+        print(language)
+        instruction = f"Answer these {questions_per_subject} questions about {subject} in {language}."
+        examples = [(question, answer) for (question, answer, _) in eleven_subjects_translated_answers[subject]]
+        while len(examples) < questions_per_subject:
+            examples += list(generate_questions_and_answers(OpenAIAPI('text-davinci-003'),
+                             instruction, examples, language_codes[language]))
 
-    subject_questions_and_answers[subject] = examples
-#%%
-subject_qa_file_path = os.path.join(reward_models_data_dir, "subject_questions_and_answers.json")
-with open(subject_qa_file_path, "w") as f:
-    json.dump(subject_questions_and_answers, f)
-#%%
+        subject_questions_and_answers[subject] = examples
+    with open(initial_example_answers_path, "w") as f:
+        json.dump(subject_questions_and_answers, f)
+else:
+    with open(initial_example_answers_path, "r") as f:
+        subject_questions_and_answers = json.load(f)
+# %%
+final_example_answers_path = os.path.join(reward_models_data_dir, "final_subject_questions_and_answers.json")
+if not os.path.exists(final_example_answers_path):
+    questions_per_subject = 100
+    responses_per_few_shot_prompt = 10
+    for language, subject in tqdm(list(zip(top_eleven_languages.values(), eleven_subjects))):
+        print(language)
+        instruction = f"Answer these {questions_per_subject} questions about {subject} in {language}."
+        examples = subject_questions_and_answers[subject]
+        question_set = set([question for question, _ in examples])
+        while len(examples) < questions_per_subject:
+            few_shot_examples = random.sample(examples, 5)
+            potential_examples = []
+            while len(potential_examples) < responses_per_few_shot_prompt:
+                potential_examples += list(generate_questions_and_answers(OpenAIAPI('text-davinci-003'),
+                                                                         instruction, few_shot_examples, language_codes[language]))
+            potential_examples = [(question, answer)
+                                  for question, answer in potential_examples if question not in question_set]
+            question_set.update([question for question, _ in potential_examples])
+            examples += potential_examples
+            print(len(examples))
+            print(examples)
 
+        subject_questions_and_answers[subject] = examples
+    with open(final_example_answers_path, "w") as f:
+        json.dump(subject_questions_and_answers, f)
+else:
+    with open(final_example_answers_path, "r") as f:
+        subject_questions_and_answers = json.load(f)
 # %%
 for subject, questions_answers in subject_questions_and_answers.items():
     print(f"Subject: {subject}")
@@ -153,7 +182,7 @@ for subject, questions_answers in subject_questions_and_answers.items():
         print()
 
 # detect wrong language
-#%%
+# %%
 for subject, questions_answers in subject_questions_and_answers.items():
     for question, answer in questions_answers:
         if detect(answer) not in top_eleven_languages or top_eleven_languages[detect(answer)] != subject:
